Code refactor: added a function exporting decorated figure with custom name (utilized for epigenetic age plots, PCA scatter matrices, batch effect evaluation plots)
